refactor(PaymentClient): 결제 관련 외부 Client 객체 추상화
